Sort .model test classes identical to tested modules
